+++ src/main/java/com/justdo/plug/member/domain/member/dto/response/MemberInfoResponse.java
@@ -1,5 +1,6 @@
 package com.justdo.plug.member.domain.member.dto.response;
 
+import com.justdo.plug.member.domain.member.Member;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -19,4 +20,11 @@ public class MemberInfoResponse {
         this.profile_url = profile_url;
     }
 
+    public static MemberInfoResponse mapMemberToMemberInfoResponse(Member member) {
+        return MemberInfoResponse.builder()
+                .email(member.getEmail())
+                .nickname(member.getNickname())
+                .profile_url(member.getProfile_url())
+                .build();
+    }
 }
