+++ src/main/java/com/justdo/plug/member/global/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.justdo.plug.member.global.utils.jwt;
+package com.justdo.plug.member.global.jwt;
 
 
 import io.jsonwebtoken.JwtException;
