+++ src/main/java/com/justdo/plug/member/global/jwt/kakao/KakaoUserInfo.java
@@ -1,4 +1,4 @@
-package com.justdo.plug.member.domain.member.dto.kakao;
+package com.justdo.plug.member.global.jwt.kakao;
 
 import lombok.Getter;
